+++ qupath-core/src/main/java/qupath/lib/objects/PathObjectTools.java
@@ -621,7 +621,7 @@ public class PathObjectTools {
 			if (includeColor)
 				pathObject.setColorRGB(null);
 		} else {
-			pathObject.setName(pathClass.getName());
+			pathObject.setName(pathClass.toString());
 			if (includeColor)
 				pathObject.setColorRGB(pathClass.getColor());				
 		}