+++ qupath-extension-processing/src/main/java/qupath/process/gui/commands/SingleMeasurementClassificationCommand.java
@@ -31,7 +31,6 @@ import javafx.scene.control.Dialog;
 import javafx.scene.control.Label;
 import javafx.scene.control.Slider;
 import javafx.scene.control.TextField;
-import javafx.scene.control.Tooltip;
 import javafx.scene.layout.GridPane;
 import javafx.stage.Modality;
 import qupath.lib.analysis.stats.Histogram;