+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## Version 0.2.0-m12
+
+This is the *release candidate* for v0.2.0 (i.e. the proposed stable version).
+
+* Several minor fixes from v0.2.0-m11, including https://github.com/qupath/qupath/issues/465
+* Major performance improvement for the 'Detect centroid distances 2D' command (by using a spatial cache)
+* Bump dependencies ImageJ, Bio-Formats
+
+
+
 ## Version 0.2.0-m11
 
 This is a *milestone* (i.e. still in development) version made available to try out new features early.